+++ core/src/main/java/com/backblaze/b2/json/FieldInfo.java
@@ -5,8 +5,6 @@
 
 package com.backblaze.b2.json;
 
-import com.backblaze.b2.util.B2Preconditions;
-
 import java.lang.reflect.Field;
 
 /**
@@ -18,6 +16,7 @@ public final class FieldInfo implements Comparable<FieldInfo> {
 
     public enum FieldRequirement { REQUIRED, OPTIONAL, IGNORED }
 
+    private final String jsonMemberName;
     public final Field field;
     public final B2JsonTypeHandler handler;
     public final FieldRequirement requirement;
@@ -28,6 +27,7 @@ public final class FieldInfo implements Comparable<FieldInfo> {
     public final boolean omitNull;
 
     /*package*/ FieldInfo(
+            String jsonMemberName,
             Field field, B2JsonTypeHandler<?> handler,
             FieldRequirement requirement,
             String defaultValueJsonOrNull,
@@ -35,6 +35,7 @@ public final class FieldInfo implements Comparable<FieldInfo> {
             boolean isSensitive,
             boolean omitNull
     ) {
+        this.jsonMemberName = jsonMemberName;
         this.field = field;
         this.handler =  handler;
         this.requirement = requirement;
@@ -46,8 +47,20 @@ public final class FieldInfo implements Comparable<FieldInfo> {
         this.field.setAccessible(true);
     }
 
+    /**
+     * Returns the member name that this field is serialized to in Json.
+     * @deprecated use {@link #getJsonMemberName()} instead which is clearer.
+     */
+    @Deprecated
     public String getName() {
-        return field.getName();
+        return jsonMemberName;
+    }
+
+    /**
+     * Returns the member name that this field is serialized to in Json.
+     */
+    public String getJsonMemberName() {
+        return jsonMemberName;
     }
 
     public B2JsonTypeHandler getHandler() {
@@ -59,7 +72,7 @@ public final class FieldInfo implements Comparable<FieldInfo> {
     }
 
     public int compareTo(@SuppressWarnings("NullableProblems") FieldInfo o) {
-        return field.getName().compareTo(o.field.getName());
+        return jsonMemberName.compareTo(o.jsonMemberName);
     }
 
     public void setConstructorArgIndex(int index) {