+++ core/src/test/java/com/backblaze/b2/json/B2JsonInferredParametersTest.java
@@ -71,6 +71,13 @@ public class B2JsonInferredParametersTest extends B2BaseTest {
                 MismatchingOrderContainer.class);
     }
 
+    @Test
+    public void testSeralizedFieldName() {
+        String json = "{\"b\": 41}";
+        final ContainerWithDifferentserializedName obj = B2Json.fromJsonOrThrowRuntime(json, ContainerWithDifferentserializedName.class);
+
+        assertEquals(41, obj.a);
+    }
 
     private static class Empty {
         @B2Json.constructor Empty() {}
@@ -175,6 +182,17 @@ public class B2JsonInferredParametersTest extends B2BaseTest {
             }
             Container other = (Container) o;
             return a == other.a && (b == null ? other.b == null : b.equals(other.b));
+        }
+    }
+
+    private static class ContainerWithDifferentserializedName {
+        @B2Json.required
+        @B2Json.serializedName(value = "b")
+        public int a;
+
+        @B2Json.constructor
+        public ContainerWithDifferentserializedName(int a) {
+            this.a = a;
         }
     }